Implemented specular mapping [description below]
Diffuse mapping and lightmapping have been added.
The diffusemap is simply the base colour of the object - the main texture -, while the specular map is a separate texture that is only used within the specular calculation of the blinn phong model and works based off a scale of gray. The whiter the map is, the more reflection it is picking up.

Also, the texture class (and the model class, consequently) has been modified to accommodate the addition of a new texture post-initialisation.